refactor: move watcher logic to new storage_watch.go and update coverage config
**Added:**

- Introduced storage_watch.go to hold fsnotify-driven manifest watcher logic, moving Watch, rewatch, watchLoop, diffAndEmit, and sendOrDrop from storage.go for better separation of concerns
- Added tests for Logger.SetRoutineID persistence and no-op behavior on empty/nil calls in session_test.go

**Changed:**

- Updated codecov.yml to ignore routine/storage_watch.go with rationale for testability and coverage expectations

**Removed:**

- Removed watcher and fsnotify event handling logic from storage.go, relocating it to storage_watch.go for modularity and maintainability
